Refactor expressions and fix minor bugs
